backend/requirements.txt
flask==3.0.2
flask-cors==4.0.0
rembg==2.0.65
pillow==10.2.0 